moved and renamed adminuserserver
